Organize train object namespaces
