+++ moorestech_client/Assets/Scripts/Client.Game/InGame/BlockSystem/PlaceSystem/TrainCar/TrainCarPlaceSystem.cs
@@ -2,7 +2,8 @@ using System.Collections.Generic;
 using System.Threading;
 using Client.Game.InGame.Context;
 using Client.Game.InGame.Train.Unit;
-using Client.Game.InGame.Train.View.Object;
+using Client.Game.InGame.Train.View.Object.Core;
+using Client.Game.InGame.Train.View.Object.Material;
 using Client.Input;
 using Cysharp.Threading.Tasks;
 using Game.Train.Unit;
