+++ moorestech_client/Assets/Scripts/Client.Game/InGame/Entity/Object/TrainCarEntityChildrenObject.cs
@@ -1,5 +1,6 @@
 using Client.Game.InGame.Context;
-using Client.Game.InGame.Train.View.Object;
+using Client.Game.InGame.Train.View.Object.Core;
+using Client.Game.InGame.Train.View.Object.Material;
 using Client.Game.InGame.UI.UIState.State;
 using UnityEngine;
