+++ moorestech_client/Assets/Scripts/Client.Game/InGame/Train/View/Object/Core/TrainCarDismountPoint.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Client.Game.InGame.Train.View.Object
+namespace Client.Game.InGame.Train.View.Object.Core
 {
     // 車両 Prefab に置き、降車地点 Transform を SerializeField で指す。
     // Placed on a train-car Prefab to point at the dismount Transform via a SerializeField.
