+++ moorestech_client/Assets/Scripts/Client.Game/InGame/Train/View/Object/Pose/TrainCarRailPositionVisualUtility.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace Client.Game.InGame.Train.View.Object
+namespace Client.Game.InGame.Train.View.Object.Pose
 {
     public static class TrainCarRailPositionVisualUtility
     {
