+++ moorestech_client/Assets/Scripts/Client.Game/InGame/UI/UIState/State/SubInventory/TrainSubInventorySource.cs
@@ -1,4 +1,4 @@
-using Client.Game.InGame.Train.View.Object;
+using Client.Game.InGame.Train.View.Object.Core;
 using Client.Game.InGame.UI.Inventory;
 using Client.Game.InGame.UI.Inventory.Train;
 using Client.Network.API;
